CON-15 - DynamoDB Docker Image
This is a basic Docker image setup for DynamoDB.  I've attached the DynamoDB binary because I think it's better than having Dockerfile downloading it from internet.  I can change it if someone feels strongly against my approach.

The DB is mounted from the host computer, but I have yet to create a started DB.  I'm not sure if mounting host folder is the best approach.  An alternative is to have a separate image with a volume where the data is stored.  The only issue is that if we ever have to rebuild the data image, the volume is kinda hard to find.  I'm open to suggestions.